Feature flag removal of discord groups (#755)
* remove old group

* rename groups-dev to groups

* replace old test with new

* handle error message
