--- groups/constants.js
@@ -1,15 +1,0 @@
-const NO_SPACES_ALLOWED = 'Roles cannot have spaces';
-const CANNOT_CONTAIN_GROUP = "Roles cannot contain 'group'.";
-const DEV_FEATURE_FLAG = 'dev';
-const SortByFields = [
-  { id: '1', fieldName: 'memberCount' },
-  { id: '2', fieldName: 'date._seconds' },
-  { id: '3', fieldName: 'lastUsedOn._seconds' },
-];
-
-export {
-  SortByFields,
-  NO_SPACES_ALLOWED,
-  CANNOT_CONTAIN_GROUP,
-  DEV_FEATURE_FLAG,
-};
